Split AppShell into focused units by concern
AppShell was a single 519-line file mixing nav-item types, the active-item
resolution logic, section/group open-state, and three chunks of presentation
(app bar, sidebar, nav list). Separate them:

- nav-items.lib.ts — the NavEntry/NavItem/Section types plus the pure
  isSection/flatten/resolveActiveItem helpers, unit-tested in isolation.
- useToggleSet.ts — the collapsible open-set state, shared by the sidebar
  sections and the nav-list groups that both duplicated it.
- AppBar / Sidebar / NavList — the presentational pieces.
- AppShell.tsx — now just composes them and owns the drawer-open state.

Behavior is unchanged; the pattern's public API (AppShell + the three nav
types) is re-exported as before.
