+++ packages/hobom-design-system/src/patterns/AppShell/index.ts
@@ -1,1 +1,2 @@
-export * from "./AppShell";
+export { AppShell } from "./AppShell";
+export type { AppShellNavItem, AppShellNavSection, NavEntry } from "./nav-items.lib";
